:hammer: add nuke confirmation

+++ src/commands/moderation/nuke.js
@@ -15,24 +15,34 @@ module.exports = {
       return message.channel.send(
         "You don't have the correct permissions for that!"
       );
-
-    let channel = client.channels.cache.get(message.channel.id);
-    const position = channel.position;
-    const topic = channel.topic;
-
-    const channel2 = await channel.clone();
-
-    channel2.setPosition(position);
-    channel2.setTopic(topic);
-    channel.delete();
-    channel2.send("Channel has been nuked! https://imgur.com/LIyGeCR");
-    await client.emit(
-      "modlog",
-      message.guild,
-      "No one",
-      "nuke",
-      "None",
-      message.member.user
+    const filter = (reaction, user) => user.id === message.author.id;
+    const Arr = ["✅", "❌"];
+    const e = await message.awaitReact(
+      "Are you sure you wanna nuke this channel?",
+      filter,
+      Arr
     );
+    if (e === "✅") {
+      let channel = client.channels.cache.get(message.channel.id);
+      const position = channel.position;
+      const topic = channel.topic;
+
+      const channel2 = await channel.clone();
+
+      channel2.setPosition(position);
+      channel2.setTopic(topic);
+      channel.delete();
+      channel2.send("Channel has been nuked! https://imgur.com/LIyGeCR");
+      await client.emit(
+        "modlog",
+        message.guild,
+        "No one",
+        "nuke",
+        "None",
+        message.member.user
+      );
+    } else {
+      return message.send("Cancelled!");
+    }
   },
 };
